+++ HibernationFixup/kern_hbfx.hpp
@@ -98,6 +98,8 @@ private:
 	bool    correct_pci_config_command {false};
 	
 	uint32_t  latestStandbyDelay {0};
+	uint32_t  latestHibernateMode {0};
+	bool      forceHibernate {false};
 	
 	/**
 	 *  Current progress mask